rename Sign in files and updated css
