+++ frontend/src/components/Header/StickyHeader.js
@@ -6,7 +6,7 @@ import {CART_RESET} from "../../constants/cartConstants";
 import {logout} from "../../actions/userActions";
 import {Link, Route} from "react-router-dom";
 import WishListTab from "../Wishlist/WishListTab";
-import SignInBtn from "../SideBar/SignInBtn";
+import SignInBtn from "../SignIn/SignInBtn";
 import logo from '../../images/bg-graphics/graphics/Final-Logo-Horizontal.png'
 
 
@@ -112,9 +112,10 @@ const StickyHeader = () => {
                                 </>
                             )}
 
-                            {/*<Nav>*/}
-                            {/*    <SignInBtn/>*/}
-                            {/*</Nav>*/}
+                        {/*     PROFILE AND LOGOUT BTNS    */}
+                            <Nav>
+                                <SignInBtn/>
+                            </Nav>
                         </>
                     )}
                 </Col>
